+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Fixed
 - Stacked Borrow friendly now.
+- `mem::Stack` capacity fixed.
 
 ## 0.9.1
 ### Changed